Extracted `_get_actor_for_job` to `common`
