+++ dashboard/modules/job/common.py
@@ -466,3 +466,15 @@ class JobDeleteResponse:
 @dataclass
 class JobLogsResponse:
     logs: str
+
+
+def _get_actor_for_job(job_id: str) -> Optional[ActorHandle]:
+    """Fetches JobSupervisor actor for job identified by Ray Job (submission) id"""
+    try:
+        return ray.get_actor(
+            JOB_ACTOR_NAME_TEMPLATE.format(job_id=job_id),
+            namespace=SUPERVISOR_ACTOR_RAY_NAMESPACE,
+        )
+    except ValueError as ve:  # Ray returns ValueError for nonexistent actor.
+        logger.warning(f"Job supervisor for job {job_id} not found: {str(ve)}")
+        return None
