+++ dashboard/modules/job/job_supervisor.py
@@ -32,10 +32,9 @@ from ray.dashboard.modules.job.common import (
     JOB_EXECUTOR_ACTOR_NAME_TEMPLATE,
     SUPERVISOR_ACTOR_RAY_NAMESPACE,
     JobInfoStorageClient,
-    JobInfo,
+    JobInfo, _get_actor_for_job,
 )
 from ray.dashboard.modules.job.job_log_storage_client import JobLogStorageClient
-from ray.dashboard.modules.job.job_manager import _get_actor_for_job
 from ray.exceptions import RuntimeEnvSetupError, ActorUnschedulableError
 from ray.job_submission import JobStatus
 from ray.runtime_env import RuntimeEnvConfig
